+++ styles/lazy-styles.css
@@ -101,4 +101,5 @@
         right: 8px;
     }
 }
+
 /* End of Cookie Pop Up Styles */
